api/requirements.txt
django>=5.2,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.6,<5.0
django-mongodb-backend>=5.2
pymongo>=4.0,<5.0
python-dotenv>=1.1,<2.0
gunicorn>=23.0,<24.0
google-generativeai>=0.8,<1.0
requests>=2.32,<3.0
slack-sdk>=3.33,<4.0
slack-bolt>=1.21,<2.0
